Issue 246 (pt.1) removed use of browser alerts, also more use of modal confirmations instead of separate pages.
