+++ app/src/main/java/org/apache/roller/weblogger/ui/struts2/editor/Categories.java
@@ -74,10 +74,6 @@ public class Categories extends UIAction {
 		return LIST;
 	}
 
-	public String move() {
-		return execute();
-	}
-
 	public List<WeblogCategory> getAllCategories() {
 		return allCategories;
 	}
