+++ app/src/main/resources/ApplicationResources.properties
@@ -40,6 +40,7 @@ generic.error.check.logs=TightBlog system error - Have administrator check Tight
 generic.changes.saved=Changes saved
 generic.toggle=Toggle
 generic.view=View
+generic.confirm=Confirm
 generic.date.toStringFormat={0,date,MM/dd/yy hh:mma}
 generic.lastModified=Last Modified
 
@@ -211,9 +212,6 @@ commentManagement.query=Filter
 commentManagement.endDate=End date
 commentManagement.startDate=Start date
 
-commentManagement.bulkDeletePrompt1=Your query matched {0} comments, 
-commentManagement.bulkDeletePrompt2=delete them all?
-commentManagement.confirmBulkDelete=Are you sure you want to delete all {0} selected by your query?
 commentManagement.lookupError=Error looking up comments
 commentManagement.pendingStatus=Pending status
 commentManagement.onlyPending=Pending only
@@ -745,9 +743,6 @@ memberPermissions.oneAdminRequired=Blog must have at least one admin.
 memberPermissions.membersRemoved=Removed {0} user(s)
 memberPermissions.membersChanged=Changed permission for {0} user(s)
 
-memberPermissions.confirmRemove=Are you sure you want to remove members from this \
-weblog?
-
 memberPermissions.saveError=Error saving permissions
 
 memberPermissions.inviteMember=Invite new member
@@ -787,7 +782,6 @@ templates.path=Path
 templates.addNewPage=Add a new template
 templates.add=Add
 templates.deleteselected=Delete Selected
-templateRemoves.title.removeOK=Confirm remove of template
 templates.error.nameAlreadyExists=A template with the name {0} already exists -- please choose another name.
 templates.error.nameSize=Template name is longer than maximum 255 characters.
 
@@ -813,7 +807,6 @@ templateEdit.save.success=Template updated successfully.
 templateEdit.launch=launch
 templateEdit.templateLanguage=Template Language
 
-templateRemoves.subtitle=Confirm removal of pages
 templateRemoves.youSure=Are you sure you want to remove these pages?
 templateRemoves.youSureWarning=<b>WARNING</b>\: Deletions are <b>NOT</b> reversible, make sure your blog does not rely \
   on these templates (or can rely on the theme defaults available, for those templates which are overrides.)
@@ -885,7 +878,6 @@ pingTargetAdd.subtitle=Add ping target
 pingTarget.pingUrl=Ping URL
 pingTarget.autoEnabled=Auto Ping?
 pingTarget.remove=Remove
-pingTarget.confirmRemoveTitle=Remove Ping Target Confirmation
 pingTarget.confirmCommonRemove=Removing this target will also remove all user''s \
 automatic ping configurations that use this target.  Are you sure?
 pingTarget.enable=Enable
@@ -926,6 +918,7 @@ planetSubscriptions.existingTitle=Existing Subscriptions
 planetSubscriptions.column.title=Title
 planetSubscriptions.column.feedUrl=Feed URL
 planetSubscription.feedUrl=Newsfeed URL
+planetSubscriptions.delete.confirm=Delete subscription from this planet?
 planetSubscription.success.deleted=Subscription successfully deleted
 planetSubscription.success.saved=Saved subscription
 planetSubscription.error=Must specify newfeed URL
