+++ app/src/main/resources/ApplicationResources_de.properties
@@ -126,8 +126,6 @@ comment.validator.excessSizeMessage=Der Kommentar hat mehr als {0} Zeichen
 comment.validator.excessSizeName=Gr\u00F6\u00DFen\u00FCberschreitung Kommentarpr\u00FCfung
 comment.validator.trackbackLinkbackMessage=Trackback von Webseite, der nicht auf Ihren Weblogeintrag verlinkt
 comment.validator.trackbackLinkbackName=Trackback R\u00FCcklink Kommentarpr\u00FCfung
-commentManagement.bulkDeletePrompt1=Ihre Abfrage lieferte {0} Kommentare,
-commentManagement.bulkDeletePrompt2=Alle l\u00F6schen?
 commentManagement.columnApproved=Zugelassen
 commentManagement.columnComment=Kommentar
 commentManagement.columnSpam=Spam
@@ -135,7 +133,6 @@ commentManagement.commentBy=Eingetragen von
 commentManagement.commentByBoth={0} (<a href="mailto:{1}">{2}</a>) von IP Adresse {3}
 commentManagement.commentByIP=Benutzer von IP Adresse {0}
 commentManagement.commentByName={0} von IP Adresse {1}
-commentManagement.confirmBulkDelete=Sind Sie sich, dass Sie alle  {0} von Ihrer Abfrage ausgew\u00E4hlten Kommentare l\u00F6schen wollen?
 commentManagement.endDate=Enddatum
 commentManagement.entry.subtitle=Kommentare managen f\u00FCr Weblogeintrag <span>{0}</span>
 commentManagement.entryTitled=Betreffend
@@ -245,7 +242,6 @@ createWebsite.tip.timezone=Die Zeitzone, die f\u00FCr die Anzeige des Weblogs ve
 createWebsite.title=Weblog anlegen
 createWebsite.weblogUrl=URL
 # ----------------------------------------------------------- Auto-installation
-templateRemoves.title.removeOK=Bitte best\u00E4tigen Sie die L\u00F6schung der Vorlage
 # --------------------------------------------------------------- Entry editors
 editor.text.name=Texteditor
 editor.xinha.name=Rich Text Editor (Xinha)
@@ -424,7 +420,6 @@ maintenance.subtitle=Weblog-Wartungsoperationen durchf\u00FChren
 maintenance.title=Webseitenwartung
 memberPermissions.administrator=Admin
 memberPermissions.author=Autor
-memberPermissions.confirmRemove=Sind Sie sicher, dass Sie Mitglieder aus diesem Weblog entfernen wollen?
 memberPermissions.description=Sie k\u00F6nnen die Rechte der Weblog-Mitglieder \u00E4ndern und oder die Benutzer l\u00F6schen indem Sie die Verwaltungsoberfl\u00E4che in der unten folgenden Tabelle verwenden und abschlie\u00DFend Speichern anklicken um die \u00C4nderungen umzusetzen. Bedenken Sie, dass Sie sich nicht selbst degradieren oder l\u00F6schen k\u00F6nnen.
 memberPermissions.inviteMember=Mitglieder einladen
 memberPermissions.limited=Beschr\u00E4nkt
@@ -477,7 +472,6 @@ ping.unknownHost=Der Hostname in der Ping Ziel URL ist unbekannt.
 pingTarget.auto=Automatisch
 pingTarget.autoEnabled=Auto Ping?
 pingTarget.confirmCommonRemove=Das L\u00F6schen dieses Ziels l\u00F6scht auch alle automatischen Benutzer Ping Konfigurationen, die dies Ziel benutzen. Sind Sie sicher?
-pingTarget.confirmRemoveTitle=L\u00F6sche Ping Ziel Best\u00E4tigung
 pingTarget.disable=Ausschalten
 pingTarget.disabled=AUS
 pingTarget.enable=Einschalten
@@ -818,5 +812,4 @@ yourWebsites.todaysHits=Heutige Hits
 yourWebsites.userCount=Mitglieder
 yourWebsites.weblog=Link
 yourWebsites.youAreInvited=Sie wurden in den Weblog [{0}] eingeladen -
-templateRemoves.subtitle=L\u00F6schung der Seiten best\u00E4tigen
 templateRemoves.youSure=Sind Sie sicher, dass Sie diese Seiten l\u00F6schen wollen?
