+++ app/src/main/resources/ApplicationResources_es.properties
@@ -257,7 +257,6 @@ memberPermissions.remove=Eliminar
 memberPermissions.inviteMember=Invitar a un nuevo miembro
 memberPermissions.membersRemoved={0} usuario(s) eliminado(s)
 memberPermissions.membersChanged=Permiso cambiado para {0} usuario(s)
-memberPermissions.confirmRemove=\u00BFEst\u00E1 seguro de que desea eliminar miembros de este weblog?
 memberPermissions.whyInvite=Invitar a alguien para ayudar a escribir el weblog.
 memberPermissions.permissionsHelpTitle=\u00BFQu\u00E9 son los permisos?
 memberPermissions.permissionHelp=Los permisos de <b>administrador</b> permiten a los usuarios enviar entradas, administrar el weblog y sus usuarios<br />Los permisos de <b>autor</b> permiten a los usuarios enviar entradas, pero no administrar el weblog<br />Los permisos <b>limitados</b> permiten a los usuarios guardar entradas solamente de tipo borrador.
@@ -272,7 +271,6 @@ templates.subtitle=Administrar plantillas que definen el weblog <span>{0}</span>
 templates.tip=Puede editar las plantillas que definan la composici\u00F3n, colores y fuentes del weblog. Tambi\u00E9n puede crear nuevas plantillas para definir p\u00E1ginas completamente nuevas. Esto es s\u00F3lo para usuarios avanzados, si no controla HTML, tal vez prefiera dejar quietas las plantillas.
 templates.addNewPage=A\u00F1adir una nueva plantilla
 templates.add=A\u00F1adir
-templateRemoves.title.removeOK=Confirmar la eliminaci\u00F3n de la plantilla
 templateEdit.path=Enlace
 templateEdit.template=Plantilla
 templateEdit.save.success=Plantilla actualizada con \u00E9xito.
@@ -294,7 +292,6 @@ ping.networkConnectionFailed=Hay problemas de conexi\u00F3n para alcanzar el sit
 ping.pingProcessingIsSuspended=El proceso ping se ha suspendido temporalmente por un administrador de sitio.
 pingTarget.pingUrl=URL para ping
 pingTarget.remove=Eliminar
-pingTarget.confirmRemoveTitle=Eliminar confirmaci\u00F3n de objetivo ping
 pingTarget.confirmCommonRemove=Eliminando este objetivo tambi\u00E9n eliminar\u00E1 la configuraci\u00F3n de pings atom\u00E1ticos de todos los usuarios que usan este objetivo. \u00BFEst\u00E1 seguro?
 pingTarget.enable=Habilitar
 pingTarget.enabled=ACTIVADO
