+++ app/src/main/resources/ApplicationResources_fr.properties
@@ -171,11 +171,6 @@ commentManagement.query=Requête
 commentManagement.endDate=Date de fin
 commentManagement.startDate=Date de début
 
-commentManagement.bulkDeletePrompt1=Votre requête a retournée 0} commentaires,
-commentManagement.bulkDeletePrompt2=Effacer tous les commentaires ?
-commentManagement.confirmBulkDelete=Etes vous certain(e) de vouloir effacer les {0} \
-commentaires retournés par votre requête ?
-
 commentManagement.pendingStatus=En attente
 commentManagement.onlyPending=En attente seulement
 commentManagement.onlyApproved=Approuvé seulement
@@ -511,8 +506,6 @@ memberPermissions.remove=Effacer
 memberPermissions.membersRemoved={0} utilisateurs enlevés.
 memberPermissions.membersChanged=Les permissions d'accés de {0} utilisateur(s) ont été changées.
 
-memberPermissions.confirmRemove=Etes vous certain(e) de vouloir enlever des membres de ce blog ?
-
 memberPermissions.inviteMember=Inviter un nouveau membre
 memberPermissions.whyInvite=Inviter quelqu'un à joindre votre blog et vous aider à écrire.
 
@@ -539,8 +532,6 @@ de les modifier.
 templates.role=Action
 templates.addNewPage=Ajouter un modèle
 templates.add=Ajouter
-templateRemoves.title.removeOK=Voulez vous vraiment effacer le modèle ?
-
 templates.error.nameAlreadyExists=Le modèle {0} existe déjà.
 
 # -------------------------------------------------------------------- Page edit
@@ -607,7 +598,6 @@ l'administrateur de ce site.
 pingTarget.pingUrl=Adresse URL
 pingTarget.autoEnabled=Ping Automatique
 pingTarget.remove=Effacer
-pingTarget.confirmRemoveTitle=Confirmation d'effacement de l'adresse de notification
 pingTarget.confirmCommonRemove=Effacer cette adresse affectera tous les blogs l'utilisant. \
 Etes vous certain de vouloir effacer celle-ci ?
