+++ app/src/main/resources/ApplicationResources_ja.properties
@@ -421,7 +421,6 @@ memberPermissions.remove=\u524A\u9664
 memberPermissions.membersRemoved={0}\u4EBA\u306E\u30E6\u30FC\u30B6\u3092\u524A\u9664\u3057\u307E\u3057\u305F
 memberPermissions.membersChanged={0}\u4EBA\u306E\u30E6\u30FC\u30B6\u6A29\u9650\u3092\u5909\u66F4\u3057\u307E\u3057\u305F
 
-memberPermissions.confirmRemove=\u3053\u306E\u30D6\u30ED\u30B0\u304B\u3089\u672C\u5F53\u306B\u524A\u9664\u3057\u3066\u3082\u3088\u308D\u3057\u3044\u3067\u3059\u304B\uFF1F
 memberPermissions.inviteMember=\u65B0\u3057\u3044\u30E6\u30FC\u30B6\u3092\u62DB\u5F85\u3059\u308B
 memberPermissions.whyInvite=\u30D6\u30ED\u30B0\u57F7\u7B46\u3092\u624B\u52A9\u3051\u3057\u3066\u304F\u308C\u308B\u4EBA\u3092\u62DB\u5F85\u3057\u307E\u3059
 
@@ -451,9 +450,6 @@ templates.tip=\u30D6\u30ED\u30B0\u306E\u30EC\u30A4\u30A2\u30A6\u30C8\u30FB\u914D
 
 templates.addNewPage=\u65B0\u898F\u30C6\u30F3\u30D7\u30EC\u30FC\u30C8\u3092\u8FFD\u52A0
 templates.add=\u8FFD\u52A0
-templateRemoves.title.removeOK=\u30C6\u30F3\u30D7\u30EC\u30FC\u30C8\u306E\u524A\u9664\u3092\u78BA\u8A8D
-
-
 # -------------------------------------------------------------------- Page edit
 
 templateEdit.path=\u30EA\u30F3\u30AF
@@ -487,7 +483,6 @@ ping.pingProcessingIsSuspended=Ping\u51E6\u7406\u306F\u30B5\u30A4\u30C8\u7BA1\u7
 pingTarget.pingUrl=Ping URL
 pingTarget.autoEnabled=\u81EA\u52D5Ping\u8A2D\u5B9A
 pingTarget.remove=\u524A\u9664
-pingTarget.confirmRemoveTitle=Ping\u30BF\u30FC\u30B2\u30C3\u30C8\u524A\u9664
 pingTarget.confirmCommonRemove=\u30BF\u30FC\u30B2\u30C3\u30C8\u3092\u524A\u9664\u3059\u308B\u3068\u3001\u3053\u306E\u30BF\u30FC\u30B2\u30C3\u30C8\u306B\u5BFE\u3059\u308B\u5168\u3066\u306E\u30E6\u30FC\u30B6\u306E\u81EA\u52D5Ping\u8A2D\u5B9A\u3082\u524A\u9664\u3055\u308C\u307E\u3059\u3002\u3088\u308D\u3057\u3044\u3067\u3059\u304B\uFF1F
 pingTarget.enable=\u30AA\u30F3\u306B\u3059\u308B
 pingTarget.enabled=\u30AA\u30F3
@@ -861,7 +856,6 @@ comment.validator.akismetMessage=Akismet(akismet.com)\u304C\u30B3\u30E1\u30F3\u3
 configForm.commentHtmlAllowed=\u30B3\u30E1\u30F3\u30C8\u3067HTML\u3092\u8A31\u53EF
 installer.whatHappenedUnknown=TightBlog\u304c\u30c7\u30fc\u30bf\u30d9\u30fc\u30b9\u3092\u78ba\u8a8d\u3057\u3066\u3044\u308b\u9593\u304b\u3001\u8d77\u52d5\u4e2d\u306e\u3044\u305a\u308c\u304b\u306b\u3001\u4e0d\u660e\u304b\u3064\u4e88\u671f\u305b\u306c\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f\u3002TightBlog\u306b\u306f\u4f55\u304c\u8d77\u3053\u3063\u305f\u306e\u304b\u7279\u5b9a\u3067\u304d\u306a\u3044\u305f\u3081\u3001\u30b5\u30fc\u30d0\u306e\u30ed\u30b0\u30fb\u30d5\u30a1\u30a4\u30eb\u3092\u78ba\u8a8d\u3057\u3066\u3001\u3042\u306a\u305f\u3054\u81ea\u8eab\u3067\u8a3a\u65ad\u3057\u306a\u3051\u308c\u3070\u306a\u308a\u307e\u305b\u3093\u3002
 installer.unknownError=\u4E0D\u660E\u306A\u30A8\u30E9\u30FC\u304C\u767A\u751F\u3057\u307E\u3057\u305F\u3002
-commentManagement.confirmBulkDelete=\u9078\u629E\u3057\u305F\u9805\u76EE{0}\u500B\u3092\u3059\u3079\u3066\u524A\u9664\u3057\u307E\u3059\u304B\uFF1F
 installer.bannerTitleRight=\u81EA\u52D5\u30A4\u30F3\u30B9\u30C8\u30FC\u30E9
 comment.validator.blacklistName=\u30D6\u30E9\u30C3\u30AF\u30EA\u30B9\u30C8\u30FB\u30B3\u30E1\u30F3\u30C8\u30FB\u30D0\u30EA\u30C7\u30FC\u30BF
 installer.cannotConnectToDatabase=\u30C7\u30FC\u30BF\u30D9\u30FC\u30B9\u306B\u63A5\u7D9A\u3067\u304D\u307E\u305B\u3093
@@ -871,7 +865,6 @@ comment.validator.excessSizeMessage=\u30B3\u30E1\u30F3\u30C8\u304C{0}\u6587\u5B5
 comment.validator.excessLinksMessage=\u30B3\u30E1\u30F3\u30C8\u306B{0}\u500B\u4EE5\u4E0A\u306E\u30EA\u30F3\u30AF\u304C\u542B\u307E\u308C\u3066\u3044\u307E\u3059
 Entry.error.textNull=\u672C\u6587\u306F\u5FC5\u9808\u9805\u76EE\u3067\u3059
 bookmarksForm.delete.confirm=\u9078\u629E\u3057\u305F\u30D6\u30C3\u30AF\u30DE\u30FC\u30AF\u3092\u524A\u9664\u3057\u307E\u3059\u304B\uFF1F
-commentManagement.bulkDeletePrompt2=\u3053\u308C\u3089\u3059\u3079\u3066\u3092\u524A\u9664\u3057\u307E\u3059\u304B\uFF1F
 tabbedmenu.design=\u30C7\u30B6\u30A4\u30F3
 configForm.ignoreSpamComments=\u30B9\u30D1\u30E0\u3068\u601D\u308F\u308C\u308B\u30B3\u30E1\u30F3\u30C8\u306F\u4FDD\u5B58\u3057\u306A\u3044
 configForm.ignoreSpamTrackbacks=\u30B9\u30D1\u30E0\u3068\u601D\u308F\u308C\u308B\u30C8\u30E9\u30C3\u30AF\u30D0\u30C3\u30AF\u306F\u4FDD\u5B58\u3057\u306A\u3044
@@ -916,7 +909,6 @@ comment.validator.excessLinksName=\u30B3\u30E1\u30F3\u30C8\u306E\u30EA\u30F3\u30
 comment.validator.excessSizeName=\u30B3\u30E1\u30F3\u30C8\u306E\u9577\u3055\u30D0\u30EA\u30C7\u30FC\u30BF
 comment.validator.trackbackLinkbackName=\u30C8\u30E9\u30C3\u30AF\u30D0\u30C3\u30AF\u30FB\u30EA\u30F3\u30AF\u30D0\u30C3\u30AF\u30B3\u30E1\u30F3\u30C8\u30D0\u30EA\u30C7\u30FC\u30BF
 comment.validator.trackbackLinkbackMessage=\u30D6\u30ED\u30B0\u30A8\u30F3\u30C8\u30EA\u3078\u30EA\u30F3\u30AF\u3055\u308C\u3066\u3044\u306A\u3044\u30B5\u30A4\u30C8\u306A\u3044\u3057\u30DA\u30FC\u30B8\u304B\u3089\u306E\u30C8\u30E9\u30C3\u30AF\u30D0\u30C3\u30AF
-commentManagement.bulkDeletePrompt1=\u30AF\u30A8\u30EA\u306B\u8A72\u5F53\u3059\u308B{0}\u306E\u30B3\u30E1\u30F3\u30C8\u304C\u898B\u3064\u304B\u308A\u307E\u3057\u305F\u3002
 commentManagement.readmore=\u3059\u3079\u3066\u306E\u30B3\u30E1\u30F3\u30C8\u3092\u898B\u308B
 commentServlet.commentAccepted=\u30B3\u30E1\u30F3\u30C8\u304C\u66F8\u304D\u8FBC\u307E\u308C\u307E\u3057\u305F\u3002
 commentServlet.previewCommentOnly=\u3053\u308C\u306F\u30D7\u30EC\u30D3\u30E5\u30FC\u3067\u3059
@@ -1077,7 +1069,6 @@ commentManagement.commentByURL=\u6295\u7A3F\u8005\u306EURL
 mediaFileView.deleteFolder=\u30D5\u30A9\u30EB\u30C0\u3092\u524A\u9664
 bookmarkForm.error.duplicateName=\u540C\u3058\u540D\u524D\u306E\u30D6\u30C3\u30AF\u30DE\u30FC\u30AF\u304C\u30D5\u30A9\u30EB\u30C0\u5185\u306B\u65E2\u306B\u5B58\u5728\u3057\u307E\u3059
 comments.ldapAuthenticatorPassword=\u3042\u306A\u305F\u306EActive Directory\u30D1\u30B9\u30EF\u30FC\u30C9\u3092\u5165\u529B\u3057\u3066\u304F\u3060\u3055\u3044\u3002
-templateRemoves.subtitle=\u30DA\u30FC\u30B8\u524A\u9664\u306E\u78BA\u8A8D
 userAdmin.userSaved=\u30E6\u30FC\u30B6\u60C5\u5831\u304C\u4FDD\u5B58\u3055\u308C\u307E\u3057\u305F
 mediaFileEdit.subtitle=\u30E1\u30C7\u30A3\u30A2\u30FB\u30D5\u30A1\u30A4\u30EB <span>{0}</span> \u306E\u7DE8\u96C6
 weblogEdit.trackbackError404=\u30C8\u30E9\u30C3\u30AF\u30D0\u30C3\u30AFURL\u306B\u5230\u9054\u3067\u304D\u306A\u304B\u3063\u305F\u305F\u3081\u30C8\u30E9\u30C3\u30AF\u30D0\u30C3\u30AF\u306B\u5931\u6557\u3057\u307E\u3057\u305F\u3002URL\u306F\u6B63\u3057\u3044\u3067\u3059\u304B?
