+++ app/src/main/resources/ApplicationResources_ko.properties
@@ -175,9 +175,6 @@ commentManagement.query=\uc9c8\uc758
 commentManagement.endDate=\uc885\ub8cc\uc77c\uc790
 commentManagement.startDate=\uc2dc\uc791\uc77c\uc790
 
-commentManagement.bulkDeletePrompt1=\uadc0\ud558\uc758 \uc9c8\uc758\uac00 {0}\uac1c\uc758 \uc758\uacac\uc774 \uac80\uc0c9\ud558\uc600\uc2b5\ub2c8\ub2e4.
-commentManagement.bulkDeletePrompt2=\ubaa8\ub450 \uc0ad\uc81c\ud558\uc2dc\uaca0\uc2b5\ub2c8\uae4c?
-commentManagement.confirmBulkDelete=\uadc0\ud558\uc758 \uc9c8\uc758\ub97c \ud1b5\ud574 \uc120\ud0dd\ub41c \ubaa8\ub4e0 {0}\uac1c\uc758 \uc758\uacac\ub4e4\uc744 \uc0ad\uc81c\ud558\uc2dc\uaca0\uc2b5\ub2c8\uae4c?
 commentManagement.pendingStatus=\ub300\uae30 \uc0c1\ud0dc
 commentManagement.onlyPending=\ub300\uae30 \uc0c1\ud0dc\ub9cc
 commentManagement.onlyApproved=\uc2b9\uc778 \uc0c1\ud0dc\ub9cc
@@ -563,8 +560,6 @@ memberPermissions.remove=\uc0ad\uc81c
 memberPermissions.membersRemoved={0} \uba85\uc758 \uc0ac\uc6a9\uc790 \uc81c\uac70
 memberPermissions.membersChanged={0} \uba85\uc758 \uc0ac\uc6a9\uc790 \uad8c\ud55c \ubcc0\uacbd
 
-memberPermissions.confirmRemove=\uc774 \uc6f9\ub85c\uadf8\uc5d0\uc11c \uba64\ubc84\ub4e4\uc744 \uc81c\uac70\ud558\uc2dc\uaca0\uc2b5\ub2c8\uae4c?
-
 memberPermissions.inviteMember=\uc0c8 \uba64\ubc84 \ucd08\ub300
 memberPermissions.whyInvite=\uc6f9\ub85c\uadf8 \uc791\uc131\uc744 \ub3d5\uae30 \uc704\ud574 \ub2e4\ub978 \uc0ac\ub78c \ucd08\ub300
 
@@ -598,8 +593,6 @@ templates.tip=\uc6f9\ub85c\uadf8\uc758 \ub808\uc774\uc544\uc6c3, \uc0c9\uae54, \
 templates.role=\uc791\uc5c5
 templates.addNewPage=\uc0c8 \ud15c\ud50c\ub9bf \ucd94\uac00
 templates.add=\ucd94\uac00
-templateRemoves.title.removeOK=\ud15c\ud50c\ub9bf\uc744 \uc0ad\uc81c\ud558\uc2dc\uaca0\uc2b5\ub2c8\uae4c?
-
 templates.error.nameAlreadyExists=\ud15c\ud50c\ub9bf, {0}\uc774(\uac00) \uc774\ubbf8 \uc874\uc7ac\ud569\ub2c8\ub2e4.
 
 # -------------------------------------------------------------------- Page edit
@@ -651,7 +644,6 @@ ping.pingProcessingIsSuspended=\uc5f0\uacb0\ud655\uc778 \ucc98\ub9ac\uac00 \uc0a
 pingTarget.pingUrl=\uc5f0\uacb0\ud655\uc778 URL
 pingTarget.autoEnabled=\uc790\ub3d9 \uc5f0\uacb0\ud655\uc778?
 pingTarget.remove=\uc0ad\uc81c
-pingTarget.confirmRemoveTitle=\uc5f0\uacb0\ud655\uc778 \ub300\uc0c1 \uc0ad\uc81c \ud655\uc778
 pingTarget.confirmCommonRemove=\uc774 \uc5f0\uacb0\ud655\uc778 \ub300\uc0c1\uc744 \uc0ad\uc81c\ud558\uac8c \ub418\uba74, \
 \uc774 \uc5f0\uacb0\ud655\uc778 \ub300\uc0c1\uc744 \uc0ac\uc6a9\ud558\ub294 \ubaa8\ub4e0 \uc0ac\uc6a9\uc790\ub4e4\uc758 \uc790\ub3d9 \uc5f0\uacb0\ud655\uc778 \uc124\uc815\ub4e4\ub3c4 \ud568\uaed8 \uc0ad\uc81c\ub429\ub2c8\ub2e4.
 pingTarget.enable=\ud65c\uc131\ud654
