+++ app/src/main/resources/ApplicationResources_ru.properties
@@ -341,7 +341,6 @@ memberPermissions.inviteMember=\u041F\u0440\u0438\u0433\u043B\u0430\u0441\u0438\
 
 memberPermissions.membersRemoved=\u0423\u0434\u0430\u043B\u0435\u043D\u043D\u044B\u0439 (-\u044B\u0435) {0} \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044C (-\u043B\u0438)
 memberPermissions.membersChanged=\u0418\u0437\u043C\u0435\u043D\u0435\u043D\u043D\u044B\u0435 \u043F\u0440\u0430\u0432\u0430 \u0434\u043B\u044F {0}\u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F (-\u043B\u0435\u0439)
-memberPermissions.confirmRemove=\u0412\u044B \u0443\u0432\u0435\u0440\u0435\u043D\u044B, \u0447\u0442\u043E \u0445\u043E\u0442\u0438\u0442\u0435 \u0443\u0434\u0430\u043B\u0438\u0442\u044C \u0447\u043B\u0435\u043D\u043E\u0432 \u0438\u0437 \u044D\u0442\u043E\u0433\u043E \u0434\u043D\u0435\u0432\u043D\u0438\u043A\u0430?
 memberPermissions.whyInvite=\u041F\u0440\u0438\u0433\u043B\u0430\u0441\u0438\u0442\u044C "\u043F\u043E\u043C\u043E\u0449\u043D\u0438\u043A\u0430" \u0434\u043B\u044F \u0432\u0435\u0434\u0435\u043D\u0438\u044F \u0441\u0432\u043E\u0435\u0433\u043E \u0434\u043D\u0435\u0432\u043D\u0438\u043A\u0430.
 
 memberPermissions.permissionsHelpTitle=\u0427\u0442\u043E \u0442\u0430\u043A\u043E\u0435 "\u043F\u0440\u0430\u0432\u0430"?
@@ -371,9 +370,6 @@ templates.tip=\u0412\u044B \u043C\u043E\u0436\u0435\u0442\u0435 \u0440\u0435\u04
 
 templates.addNewPage=\u0414\u043E\u0431\u0430\u0432\u0438\u0442\u044C \u043D\u043E\u0432\u044B\u0439 \u0448\u0430\u0431\u043B\u043E\u043D
 templates.add=Add - \u0414\u043E\u0431\u0430\u0432\u0438\u0442\u044C
-templateRemoves.title.removeOK=\u041F\u043E\u0434\u0442\u0432\u0435\u0440\u0436\u0434\u0435\u043D\u0438\u0435 \u0443\u0434\u0430\u043B\u0435\u043D\u0438\u0435 \u0448\u0430\u0431\u043B\u043E\u043D\u0430.
-
-
 # -------------------------------------------------------------------- Page edit
 
 templateEdit.path=\u0421\u0441\u044B\u043B\u043A\u0430
@@ -405,7 +401,6 @@ ping.pingProcessingIsSuspended=\u0412\u043E\u0437\u043C\u043E\u0436\u043D\u043E\
 
 pingTarget.pingUrl=URL \u043F\u0438\u043D\u0433\u0430
 pingTarget.remove=\u0423\u0434\u0430\u043B\u0438\u0442\u044C
-pingTarget.confirmRemoveTitle=\u0423\u0434\u0430\u043B\u0438\u0442\u044C \u043F\u043E\u0434\u0442\u0432\u0435\u0440\u0436\u0434\u0435\u043D\u0438\u0435 \u0430\u0434\u0440\u0435\u0441\u0430 \u043F\u0438\u043D\u0433\u0430
 pingTarget.confirmCommonRemove=\u0423\u0434\u0430\u043B\u0435\u043D\u0438\u0435 \u044D\u0442\u043E\u0433\u043E \u0430\u0434\u0440\u0435\u0441\u0430 \u043F\u043E\u0432\u043B\u0435\u0447\u0435\u0442 \u0437\u0430 \u0441\u043E\u0431\u043E\u0439 \u0443\u0434\u0430\u043B\u0435\u043D\u0438\u0435 \u0432\u0441\u0435\u0445 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044C\u0441\u043A\u0438\u0445 \u0430\u0432\u0442\u043E\u043C\u0430\u0442\u0438\u0447\u0435\u0441\u043A\u0438\u0445 \u043D\u0430\u0441\u0442\u0440\u043E\u0435\u043A \u043F\u0438\u043D\u0433\u0430, \u0438\u0441\u043F\u043E\u043B\u044C\u0437\u0443\u044E\u0449\u0438\u0445 \u044D\u0442\u043E\u0442 \u0430\u0434\u0440\u0435\u0441. \u0412\u044B \u0443\u0432\u0435\u0440\u0435\u043D\u044B?
 pingTarget.enable=Enable
 pingTarget.enabled=ON
