+++ app/src/main/resources/struts.xml
@@ -431,18 +431,6 @@
             <result name="list" type="chain">templates</result>
         </action>
         
-        <action name="templatesRemove!*" method="{1}"
-                class="org.apache.roller.weblogger.ui.struts2.editor.TemplatesRemove">
-            <result name="input" type="tiles">.TemplatesRemove</result>
-            <result name="confirm" type="tiles">.TemplatesRemove</result>
-            <result name="success" type="chain">templates</result>
-            <result name="cancel" type="redirectAction">
-                <param name="actionName">templates</param>
-                <param name="weblog">${weblog}</param>
-            </result>
-            <result name="error" type="chain">templates</result>
-        </action>
-        
         <action name="members!*" method="{1}"
                 class="org.apache.roller.weblogger.ui.struts2.editor.Members">
             <result name="list" type="tiles">.Members</result>
