+++ app/src/main/webapp/WEB-INF/jsps/editor/TemplatesSidebar.jsp
@@ -16,7 +16,6 @@
   directory of this distribution.
 --%>
 <%@ include file="/WEB-INF/jsps/taglibs-struts2.jsp" %>
-<script src="<s:url value='/tb-ui/scripts/jquery-2.1.1.min.js' />"></script>
 <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.2.23/angular.min.js"></script>
 
 <div class="sidebarFade">
